requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alt-profanity-check @ git+https://github.com/dimitrismistriotis/alt-profanity-check.git@4d5225a8c4a38aa85ff820b3cac27c8491a4b705
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord.py==2.3.2
docopt-ng==0.9.0
emoji==2.11.1
fogleman-TWL06-scrabble @ git+https://github.com/battisti456/fogleman_TWL06_scrabble.git@96e9e91a46570e9c82439e962e965235659dac83
color_tools_battisti456 @ git+https://github.com/battisti456/color_tools_battisti456.git
config_system_battisti456 @ git+https://github.com/battisti456/config_system_battisti456.git
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
inflect==4.1.0
joblib==1.4.2
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
python-chess==1.999
python-dateutil==2.9.0.post0
Python-Trivia-API==4.0.2
pytz==2024.1
regex==2024.4.28
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
threadpoolctl==3.5.0
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wonderwords==2.2.0
word-forms==2.1.0
yarl==1.9.4